% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cel_devstages-data.R
\docType{data}
\name{Cel_devstages}
\alias{Cel_devstages}
\title{Cel_devstages}
\format{
An object of class \code{list} of length 2.
}
\usage{
Cel_devstages
}
\description{
Table of reference datasets and landmark developmental stages.
}
\details{
The timing of developmental stages were compiled from the WormBook \insertCite{wormbook}{wormRef} and the publications
associated with the datasets used to build the references.
This mainly concerns the larval reference \insertCite{hyun2013dampening}{wormRef}
}
\section{Format }{

A list with 2 dataframes:
\itemize{
\item \verb{$devstages}
\item \verb{$datasets}
}

Both dataframes have the following fields:
\itemize{
\item \code{name}
\item \code{tstart}
\item \code{tend}
\item \code{tunit}
}
}

\references{
\insertAllCited{}
}
\keyword{datasets}
