% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cel_genes-data.R
\docType{data}
\name{Cel_genes}
\alias{Cel_genes}
\title{Cel_genes}
\format{
An object of class \code{data.frame} with 61541 rows and 5 columns.
}
\source{
\href{https://parasite.wormbase.org/info/Tools/biomart.html}{paraSite biomaRt}
}
\usage{
Cel_genes
}
\description{
A gene ID table for \emph{C. elegans}.
}
\details{
This table was built directly from the paraSite biomaRt.
}
\section{Format}{

A dataframe with the folowing columns :
\itemize{
\item \code{wb_id}: WBGene ID (\emph{e.g} \code{WBGene00019636})
\item \code{transcript_name}: Transcript name (\emph{e.g} \code{K10F12.4a.1})
\item \code{sequence_name}: Sequence name (\emph{e.g} \code{K10F12.4}) - this is most frequently used in transcriptomic datasets
\item \code{public_name}: Gene name (\emph{e.g} \code{gsto-3})
\item \code{transcript_length}: Transcript length from 5'UTR start to 3'UTR end (\emph{e.g} \code{3161})
}
}

\keyword{datasets}
