% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cel_YA_1-data.R
\docType{data}
\name{Cel_YA_1}
\alias{Cel_YA_1}
\title{Cel_YA_1}
\format{An object of class \code{list} of length 3.}
\source{
\href{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-7573/}{E-MTAB-7573}
}
\usage{
Cel_YA_1
}
\description{
Transcriptomic data from \emph{C. elegans} Young adult samples.
}
\details{
One time series is included in this dataset.
\itemize{
\item \insertCite{van2019balancing}{wormRef}
}

The gene expression matrix is \eqn{log(X + 1)} of quantile-normalized raw data.
}
\section{Format }{

A list with :
\itemize{
\item \verb{$g}: The gene expression matrix (genes as rows, samples as columns)
\item \verb{$p}: A dataframe of phenotypic data on the samples, usually :
\itemize{
\item \code{sname} : The sample names (and colnames of the gene expression matrix).
\item \code{age} : Sample developmental age. May be scaled to join multiple datasets.
\item \code{cov} : A factor separating the samples into groups if needed (\emph{e.g} multiple datasets, batches); else NULL.
\item \code{age_ini} : Sample chronological age, as given in the literature.
\item \code{accession} : Sample accession ID for GEO/ArrayExpress.
}
\item \verb{$geim_params}: A list of elements needed for the gene expression interpolation model :
\itemize{
\item \verb{$formula} : the model formula.
\item \verb{$method} : the model type (eg. gam, glm).
\item \verb{$dim_red} : the dimension reduction method to use (pca, ica).
\item \verb{$nc}: The number PCA/ICA components to use in the model.
}
}
}

\references{
\insertAllCited{}
}
\keyword{datasets}
