% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_refs.R
\name{plot_refs_Cel}
\alias{plot_refs_Cel}
\title{Plot Reference Timelines}
\usage{
plot_refs_Cel()
}
\description{
Plots the coverage of the \code{wormRef} references (requires ggplot2)
}
\examples{
\donttest{
plot_refs_Cel()
}

}
