% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cel_embryo-data.R
\docType{data}
\name{Cel_embryo}
\alias{Cel_embryo}
\title{Cel_embryo}
\format{An object of class \code{list} of length 3.}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE50548}{GSE50548}
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60755}{GSE60755}
}
\usage{
Cel_embryo
}
\description{
Transcriptomic data from \emph{C. elegans} embryonic samples.
}
\details{
Two time series are included in this dataset.
\itemize{
\item \code{H}: \insertCite{hashimshony2015spatiotemporal}{wormRef}
\item \code{L}: \insertCite{levin2016mid}{wormRef}
}

The gene expression matrix is \eqn{log(X + 1)} of quantile-normalized raw data.
}
\section{Format }{

A list with :
\itemize{
\item \verb{$g}: The gene expression matrix (genes as rows, samples as columns)
\item \verb{$p}: A dataframe of phenotypic data on the samples, usually :
\itemize{
\item \code{sname} : The sample names (and colnames of the gene expression matrix).
\item \code{age} : Sample developmental age. May be scaled to join multiple datasets.
\item \code{cov} : A factor separating the samples into groups if needed (\emph{e.g} multiple datasets, batches); else NULL.
\item \code{age_ini} : Sample chronological age, as given in the literature.
\item \code{accession} : Sample accession ID for GEO/ArrayExpress.
}
\item \verb{$df}: The degree of freedom to use for the PLSR interpolation model.
\item \verb{$nc}: The number of PLS components to use for the prediction of the PLSR interpolation model
}
}

\references{
\insertAllCited{}
}
\keyword{datasets}
