% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepref_Cel.R
\name{Cel_prep}
\alias{Cel_prep}
\alias{.prepref_Cel_embryo}
\alias{.prepref_Cel_larval}
\alias{.prepref_Cel_YA_1}
\alias{.prepref_Cel_YA_2}
\title{Build Interpolated Gene Expression Reference}
\usage{
.prepref_Cel_embryo(n.inter)

.prepref_Cel_larval(n.inter)

.prepref_Cel_YA_1(n.inter)

.prepref_Cel_YA_2(n.inter)
}
\arguments{
\item{n.inter}{passed on to \code{\link[RAPToR]{plsr_interpol}}}
}
\value{
A list with \code{interpGE} the interpolated gene expression matrix and
\code{time.series} the time of the interpGE matrix columns.
}
\description{
Builds the PLSR interpolation of the reference dataset.
\strong{These functions are internally called by \code{\link[RAPToR]{prepare_refdata}} from RAPToR.}
}
\seealso{
\code{\link[RAPToR]{plsr_interpol}} \code{\link[RAPToR]{ge_im}}
}
