% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepref_Cel.R
\name{Cel_prep}
\alias{Cel_prep}
\alias{.prepref_skel}
\alias{.prepref_Cel_embryo}
\alias{.prepref_Cel_larval}
\alias{.prepref_Cel_larv_YA}
\alias{.prepref_Cel_YA_1}
\alias{.prepref_Cel_YA_2}
\alias{.prepref_Cel_aging_1}
\title{Build Interpolated Gene Expression References}
\usage{
.prepref_skel(data, from = NULL, to = NULL)

.prepref_Cel_embryo(n.inter = NULL, by.inter = NULL)

.prepref_Cel_larval(n.inter = NULL, by.inter = NULL)

.prepref_Cel_larv_YA(n.inter = NULL, by.inter = NULL)

.prepref_Cel_YA_1(n.inter = NULL, by.inter = NULL)

.prepref_Cel_YA_2(n.inter = NULL, by.inter = NULL)

.prepref_Cel_aging_1(n.inter = NULL, by.inter = NULL)
}
\arguments{
\item{n.inter, by.inter}{the resolution of the interpolation, as in \code{seq(start, end, length.out = n.inter)} or \code{seq(start, end, by = by.inter)}.}
}
\value{
A list with \code{interpGE} the interpolated gene expression matrix and
\code{time.series} the time of the interpGE matrix columns.
}
\description{
Builds the interpolation of the reference datasets.
\strong{These functions are internally called by \code{\link[RAPToR]{prepare_refdata}} from RAPToR.}
}
\seealso{
\code{\link[RAPToR]{prepare_refdata}} \code{\link[RAPToR]{ge_im}} \code{\link[RAPToR]{make_ref}}
}
